+++ projects/mk-kit/core/index.ts
@@ -5,6 +5,7 @@ export * from './a11y/live-announcer.service';
 export * from './a11y/focus-trap';
 export * from './a11y/unique-id';
 export * from './overlay/overlay-ref';
+export * from './overlay/overlay-root';
 export * from './overlay/overlay.service';
 export * from './overlay/anchored-overlay';
 export * from './i18n/mk-i18n';